Fix nested arrays.
We need to allocate the slice once, set the value of the field in the
provided struct to that allocated slice, and then iteratively use the
index method on the slice reflect.Value to set the elements of the
slice. Otherwise it's not putting the value in the right heap space
and panics about an unaddressable stack value.