Fixups and pointer support.
Added support for slices containing pointers to struct.

Started cleaning up support for different int sizes in structs. These seem to
require some manual coding and shrinking to smaller types. More implementation
and exhaustive testing is needed in this regard.

We should also clean up all the type assertions to check the type and return an
error if there is a mismatch instead of panicing because it would be bad for a
server process to panic when a client sends malformed CBOR.

For the purposes of using this library for RAINS, issue #8 will need to be fixed
which seems most suitable by not having interface types in the data struct, as
this is how the previous cap'n proto implementation was working. Personally I
feel that if there is only one implementation of that interface there is not
much benefit for it being an interface, it is much easier to just have a struct
which the needed methods are implemented on. If replacement is a concern then
it could be placed in a different package which can be swapped out easily.