Support tagged structs in interfaces.
This commit finally adds support for interface types in structures which are
being unmarshaled. The user must first specify a mapping for all the structs
which are desired to be put in interface typed fields of structs which will be
marshaled, or directly marshaled with a variable of type &interface. This does
not yet support interface tagged structs in slices or arrays yet.