Add support for tags in maps and arrays.
Cleaned up the edge cases to properly support tagged elements in arrays and
maps. Now the array and map methods return a tagged type, which can be untagged
by calling the appropriate method. Maps containing arrays and arrays containing
maps will be cross-untagged.

Modified the tests to untag the results of the read calls so that they can be
compared to the input structures.

Implemented interface setting on inner structs which are containined in slices.
At the moment it will set an interface slice to a slice of the concrete struct,
but it would also be possible to make a tunable option to choose between
setting a pointer and setting the concrete value in the interface.