Add support for tagged structs (#11)
* Support tagged structs in interfaces.

This commit finally adds support for interface types in structures which are
being unmarshaled. The user must first specify a mapping for all the structs
which are desired to be put in interface typed fields of structs which will be
marshaled, or directly marshaled with a variable of type &interface. This does
not yet support interface tagged structs in slices or arrays yet.

* Address review comments.

* Add support for tags in maps and arrays.

Cleaned up the edge cases to properly support tagged elements in arrays and
maps. Now the array and map methods return a tagged type, which can be untagged
by calling the appropriate method. Maps containing arrays and arrays containing
maps will be cross-untagged.

Modified the tests to untag the results of the read calls so that they can be
compared to the input structures.

Implemented interface setting on inner structs which are containined in slices.
At the moment it will set an interface slice to a slice of the concrete struct,
but it would also be possible to make a tunable option to choose between
setting a pointer and setting the concrete value in the interface.

* Update reader.go